infra/test_setup/consumer/requirements.txt
pynsq==1.0.0 